Fix AddressFormatType row block, and move it in the Locations namespace since it is not a shared form type
